% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/HW6-package.R
\docType{package}
\name{HW6}
\alias{HW6}
\alias{HW6-package}
\title{Advanced R programming homework 6}
\description{
The package contains three utility functions: mode(), perms() and shortestpath()
}

