% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mode}
\alias{mode}
\title{Mode Function}
\usage{
mode(z)
}
\arguments{
\item{z}{- a numeric vector.}
}
\value{
Mode - the most frequent number.
}
\description{
Finds mode of given numeric vector.
}

