% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{perms}
\alias{perms}
\title{Perm Function}
\usage{
perms(n)
}
\arguments{
\item{n}{a positive integer.}
}
\value{
P matrix with permutations in rows.
}
\description{
Calculates all permutations for an integer vector of given length n.
}

