% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{shortestpath}
\alias{shortestpath}
\title{Find the length of shortest path.}
\usage{
shortestpath(G, x, y)
}
\arguments{
\item{G}{an 0-1 adjacency matrix which describes graph.}

\item{x}{a vertex number from which path should be calculated.}

\item{y}{a vertex number to which path should be calculated.}
}
\value{
d Length of shortest path between two given vertices.
}
\description{
Calculates the length of shortest path.
}

